Combine ofType and ofTypes
Previously we could not figure out how to create a strictly typed
version of ofType with variable arity. The type
UnknownActionCreatorWithPayload allows us to infer the overlap of the
payload between multiple action creators, and thus combine these two
functions.

Overloads have been added to allow for usage when there is no overlap
between the payload, but all the action creators have payloads, (in
which case the overlap is `{}`), and for when at least one of the
action creators create void actions.