Make actions readonly and protected
This recursively marks all created actions (and everything in their
payloads) as `readonly`, and protects them with the use of
`Object.defineProperty` with `writable: false`.

There is a performance aspect to doing a deep copy and protecting the
properties of all created action objects, but it is measured in
microseconds. Deep copy and protecting is about half as fast as
protecting in place[1].

[1]: https://jsperf.com/deep-copy-and-protect/130