Remove runtime overhead of markers
Previously all markers would add an extra function call for each emitted
value in an observable. This was because the markers where added by
using `Observable.lift`, which leaves it to the operator to handle
emitted values.

This new implementation just extends Observable with the marker field,
and lets the Observable implementation handle subscribing to the source.
This means we don't add any extra function calls for the emitted values
to pass through.