fix: Mark rxjs as a peer dependency
We need to make sure that the users of the library use the exact same
version of rxjs as the library does. We've had a few incidents with this
in the past, so it's nice to solve properly.

The peer dependency has a bit wider range than the dev dependency, to
make it easier to adopt.