fix(state stream): ensure state property is always updated
Using `reducedState$.subscribe(this.subject)` works fine in all our tests, but when trying
to use StateStream in ardoq-front I ran into issues with the `.state` property not being
updated. I have not managed to write a test to verify this behavior, but manually creating
an Observer fixed the issue.